+++ package.json
@@ -33,6 +33,8 @@
     "react-hook-form": "^7.51.3",
     "react-lazy-load-image-component": "^1.6.0",
     "react-router-dom": "^6.22.3",
+    "react-viewport-list": "^7.1.2",
+    "react-virtualized": "^9.22.5",
     "zod": "^3.22.4"
   },
   "devDependencies": {
@@ -40,6 +42,7 @@
     "@eslint/js": "8.57.0",
     "@faker-js/faker": "^8.4.1",
     "@mswjs/data": "^0.16.1",
+    "@nabla/vite-plugin-eslint": "^2.0.3",
     "@stylistic/eslint-plugin-migrate": "^1.7.0",
     "@tanstack/eslint-plugin-query": "^5.28.11",
     "@testing-library/dom": "^10.0.0",
@@ -53,6 +56,7 @@
     "@types/react": "^18.2.78",
     "@types/react-dom": "^18.2.25",
     "@types/react-lazy-load-image-component": "^1.6.3",
+    "@types/react-virtualized": "^9",
     "@typescript-eslint/eslint-plugin": "^7.6.0",
     "@typescript-eslint/parser": "^7.6.0",
     "@vitejs/plugin-react": "^4.2.1",
