+++ src/containers/Activite/MyAnimeList/renderChild.tsx
@@ -13,3 +13,13 @@ export const cardRenderChild = (animelist: Array<Omit<MALAnime, "broadcast"> & A
       <MALCard malAnime={anime} />
     </Grid>
   ))
+
+export const singleCardRender = (anime: Omit<MALAnime, "broadcast"> & AnimeDTO, key: number) => (
+  <Grid key={anime.malId} lg={3} md={4} xs={12} sm={6}>
+    <MALCard malAnime={anime} />
+  </Grid>
+)
+
+export const singleTableRender = (anime: Omit<MALAnime, "broadcast"> & AnimeDTO, key: number) => (
+  <MALRow key={anime.malId} malAnime={anime} />
+)
