+++ src/containers/Footer/Footer.tsx
@@ -1,3 +1,4 @@
+/* global APP_VERSION */
 import Box from "@mui/material/Box"
 import Container from "@mui/material/Container"
 import Typography from "@mui/material/Typography"
