+++ src/hooks/containers/Activite/MyAnimeList/useMyAnimeList.ts
@@ -1,14 +1,9 @@
-import { useMutation, useQuery } from "@tanstack/react-query"
+import { useQuery } from "@tanstack/react-query"
 
 import { type AnimeType } from "@/enums/AnimeType"
-import StorageState from "@/enums/StorageState"
 import UserAnimeStatus from "@/enums/UserAnimeStatus"
 import { useMyAnimeListContext } from "@/hooks/context/useMyAnimeListContext"
-import { type AnimeDTO } from "@/interfaces/services/AnimeService/AnimeDTO"
-import { type AnimeInServerDTO } from "@/interfaces/services/AnimeService/AnimeInServerDTO"
-import type ResponseError from "@/interfaces/services/ResponseError"
 import type MALAnime from "@/interfaces/services/UserService/MyAnimeList/MALAnime"
-import AnimeServices from "@/services/AnimeService"
 import UserService from "@/services/UserService"
 import { useCallback, useEffect } from "react"
 import useLibrary from "../../AnimeLibrary/useLibrary"
@@ -16,7 +11,11 @@ import useLibrary from "../../AnimeLibrary/useLibrary"
 const useMyAnimeList = () => {
   const { animes: animeLibrary, updateAnime: updateAnimeInLibrary } = useLibrary()
 
-  const { myAnimeList, setMyAnimeList, setUpdateAnimeStateFunction } = useMyAnimeListContext()
+  const {
+    myAnimeList,
+    setMyAnimeList
+    // setUpdateAnimeStateFunction
+  } = useMyAnimeListContext()
 
   const { data, isFetched, isFetching } = useQuery({
     staleTime: 3600000,
@@ -72,119 +71,119 @@ const useMyAnimeList = () => {
       setMyAnimeList(data.sort(sortMALAnimeList).map(mappedMALAnime))
   }, [data, isFetched, mappedMALAnime])
 
-  // Delete Anime
-  const deleteCall = useCallback(async (defaultAnime: AnimeDTO) => {
-    await AnimeServices.delete(defaultAnime.malId)
-  }, [])
+  // // Delete Anime
+  // const deleteCall = useCallback(async (defaultAnime: AnimeDTO) => {
+  //   await AnimeServices.delete(defaultAnime.malId)
+  // }, [])
 
-  const onErrorDelete = useCallback((error: ResponseError, defaultAnime: AnimeDTO) => {
-    console.error(
-      "Une erreur est survenue lors de la supression l'anime %s avec le status %s",
-      defaultAnime.malId,
-      error?.response?.status
-    )
-  }, [])
+  // const onErrorDelete = useCallback((error: ResponseError, defaultAnime: AnimeDTO) => {
+  //   console.error(
+  //     "Une erreur est survenue lors de la supression l'anime %s avec le status %s",
+  //     defaultAnime.malId,
+  //     error?.response?.status
+  //   )
+  // }, [])
 
-  const onSuccesReset = useCallback(
-    (_: void, defaultAnime: AnimeDTO) => {
-      updateAnimeInLibrary(defaultAnime)
-      setMyAnimeList(prev =>
-        prev.map(prevAnime => (prevAnime.id === defaultAnime.malId ? { ...prevAnime, ...defaultAnime } : prevAnime))
-      )
-    },
-    [updateAnimeInLibrary]
-  )
+  // const onSuccesReset = useCallback(
+  //   (_: void, defaultAnime: AnimeDTO) => {
+  //     updateAnimeInLibrary(defaultAnime)
+  //     setMyAnimeList(prev =>
+  //       prev.map(prevAnime => (prevAnime.id === defaultAnime.malId ? { ...prevAnime, ...defaultAnime } : prevAnime))
+  //     )
+  //   },
+  //   [updateAnimeInLibrary]
+  // )
 
-  const { isPending: isDeletePending, mutate: deleteAnime } = useMutation<void, ResponseError, AnimeDTO>({
-    mutationFn: deleteCall,
-    onSuccess: onSuccesReset,
-    onError: onErrorDelete
-  })
+  // const { isPending: isDeletePending, mutate: deleteAnime } = useMutation<void, ResponseError, AnimeDTO>({
+  //   mutationFn: deleteCall,
+  //   onSuccess: onSuccesReset,
+  //   onError: onErrorDelete
+  // })
 
-  // Save Anime
-  const saveInLibraryCall = useCallback(async (malId: number) => await AnimeServices.saveInLibrary(malId), [])
+  // // Save Anime
+  // const saveInLibraryCall = useCallback(async (malId: number) => await AnimeServices.saveInLibrary(malId), [])
 
-  const onSuccessSaveInLibrary = useCallback(
-    (anime: AnimeDTO) => {
-      updateAnimeInLibrary(anime)
+  // const onSuccessSaveInLibrary = useCallback(
+  //   (anime: AnimeDTO) => {
+  //     updateAnimeInLibrary(anime)
 
-      setMyAnimeList(prev =>
-        prev.map(prevAnime => (prevAnime.id === anime.malId ? { ...prevAnime, ...anime } : prevAnime))
-      )
-    },
-    [updateAnimeInLibrary, setMyAnimeList]
-  )
+  //     setMyAnimeList(prev =>
+  //       prev.map(prevAnime => (prevAnime.id === anime.malId ? { ...prevAnime, ...anime } : prevAnime))
+  //     )
+  //   },
+  //   [updateAnimeInLibrary, setMyAnimeList]
+  // )
 
-  const onErrorSaveInLibrary = useCallback((error: ResponseError, malId: number) => {
-    console.error(
-      "Une erreur est survenue lors de l'enregistrement de l'anime %s avec le status %s",
-      malId,
-      error?.response?.status
-    )
-  }, [])
+  // const onErrorSaveInLibrary = useCallback((error: ResponseError, malId: number) => {
+  //   console.error(
+  //     "Une erreur est survenue lors de l'enregistrement de l'anime %s avec le status %s",
+  //     malId,
+  //     error?.response?.status
+  //   )
+  // }, [])
 
-  const { isPending: isSaveInLibraryPending, mutate: saveAnime } = useMutation<AnimeDTO, ResponseError, number>({
-    mutationFn: saveInLibraryCall,
-    onSuccess: onSuccessSaveInLibrary,
-    onError: onErrorSaveInLibrary
-  })
+  // const { isPending: isSaveInLibraryPending, mutate: saveAnime } = useMutation<AnimeDTO, ResponseError, number>({
+  //   mutationFn: saveInLibraryCall,
+  //   onSuccess: onSuccessSaveInLibrary,
+  //   onError: onErrorSaveInLibrary
+  // })
 
-  // Update Storage State
-  const updateStorageStateCall = useCallback(
-    async ({ storageState, malId }: { storageState: string; malId: number }) =>
-      await AnimeServices.updateStorageState(malId, storageState),
-    []
-  )
+  // // Update Storage State
+  // const updateStorageStateCall = useCallback(
+  //   async ({ storageState, malId }: { storageState: string; malId: number }) =>
+  //     await AnimeServices.updateStorageState(malId, storageState),
+  //   []
+  // )
 
-  const onErrorUpdateStorageState = useCallback((error: ResponseError, { malId }: { malId: number }) => {
-    console.error(
-      "Une erreur est survenue lors de la mise à jour du storage state de l'anime %s avec le status %s",
-      malId,
-      error?.response?.status
-    )
-  }, [])
+  // const onErrorUpdateStorageState = useCallback((error: ResponseError, { malId }: { malId: number }) => {
+  //   console.error(
+  //     "Une erreur est survenue lors de la mise à jour du storage state de l'anime %s avec le status %s",
+  //     malId,
+  //     error?.response?.status
+  //   )
+  // }, [])
 
-  const onSuccessUpdateAnimeInServer = useCallback(
-    (anime: AnimeInServerDTO) => {
-      updateAnimeInLibrary(anime)
-      setMyAnimeList(prev =>
-        prev.map(prevAnime => (prevAnime.id === anime.malId ? { ...prevAnime, ...anime } : prevAnime))
-      )
-    },
-    [updateAnimeInLibrary]
-  )
+  // const onSuccessUpdateAnimeInServer = useCallback(
+  //   (anime: AnimeInServerDTO) => {
+  //     updateAnimeInLibrary(anime)
+  //     setMyAnimeList(prev =>
+  //       prev.map(prevAnime => (prevAnime.id === anime.malId ? { ...prevAnime, ...anime } : prevAnime))
+  //     )
+  //   },
+  //   [updateAnimeInLibrary]
+  // )
 
-  const { isPending: isUpdateStorageStatePending, mutate: setStorageState } = useMutation<
-    AnimeInServerDTO,
-    ResponseError,
-    { storageState: string; malId: number }
-  >({
-    mutationFn: updateStorageStateCall,
-    onSuccess: onSuccessUpdateAnimeInServer,
-    onError: onErrorUpdateStorageState
-  })
+  // const { isPending: isUpdateStorageStatePending, mutate: setStorageState } = useMutation<
+  //   AnimeInServerDTO,
+  //   ResponseError,
+  //   { storageState: string; malId: number }
+  // >({
+  //   mutationFn: updateStorageStateCall,
+  //   onSuccess: onSuccessUpdateAnimeInServer,
+  //   onError: onErrorUpdateStorageState
+  // })
 
-  const updateAnime = useCallback(
-    (anime: AnimeDTO) => {
-      if (!isUpdateStorageStatePending && !isSaveInLibraryPending && !isDeletePending) {
-        const { storageState } = anime
-        if (storageState === undefined) {
-          saveAnime(anime.malId)
-        }
-        if (storageState === StorageState.FLUX_FROID) {
-          setStorageState({ storageState: StorageState.FLUX_CHAUD, malId: anime.malId })
-        }
-        if (storageState === StorageState.FLUX_CHAUD) {
-          deleteAnime({ ...anime, storageState: undefined, addedOnServer: undefined })
-        }
-      }
-    },
-    [isUpdateStorageStatePending, isSaveInLibraryPending, isDeletePending]
-  )
+  // const updateAnime = useCallback(
+  //   (anime: AnimeDTO) => {
+  //     if (!isUpdateStorageStatePending && !isSaveInLibraryPending && !isDeletePending) {
+  //       const { storageState } = anime
+  //       if (storageState === undefined) {
+  //         saveAnime(anime.malId)
+  //       }
+  //       if (storageState === StorageState.FLUX_FROID) {
+  //         setStorageState({ storageState: StorageState.FLUX_CHAUD, malId: anime.malId })
+  //       }
+  //       if (storageState === StorageState.FLUX_CHAUD) {
+  //         deleteAnime({ ...anime, storageState: undefined, addedOnServer: undefined })
+  //       }
+  //     }
+  //   },
+  //   [isUpdateStorageStatePending, isSaveInLibraryPending, isDeletePending]
+  // )
 
-  useEffect(() => {
-    setUpdateAnimeStateFunction(() => updateAnime)
-  }, [updateAnime])
+  // useEffect(() => {
+  //   setUpdateAnimeStateFunction(() => updateAnime)
+  // }, [updateAnime])
 
   return { myAnimeList, isFetching }
 }
