+++ src/interfaces/contexts/MyAnimeListContext.ts
@@ -16,4 +16,6 @@ export default interface MyAnimeListContext {
   setSelectedRenderMode: Dispatch<SetStateAction<RenderMode>>
   updateAnimeStateFunction: (a: AnimeDTO) => void
   setUpdateAnimeStateFunction: Dispatch<SetStateAction<(a: AnimeDTO) => void>>
+  page: number
+  setPage: Dispatch<SetStateAction<number>>
 }
